Move releases table row to separate component
